requirements.txt
amqp==5.3.1
anyio==4.10.0
asgiref==3.8.1
billiard==4.2.1
cachetools==5.5.2
celery==5.5.3
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.1
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-filter==25.1
django-formtools==2.5.1
django-otp==1.6.0
django-phonenumber-field==8.1.0
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
flower==2.0.1
google-auth==2.40.2
grpcio==1.73.1
grpcio-tools==1.73.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.12.3
idna==3.10
inflection==0.5.1
kombu==5.5.4
packaging==25.0
pillow==11.2.1
prometheus_client==0.22.1
prompt_toolkit==3.0.51
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydub==0.25.1
PyJWT==2.9.0
pypng==0.20220715.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
qrcode==7.4.2
redis==6.2.0
requests==2.32.3
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tornado==6.5.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
